vu meter audio visualization

+++ main/src/audioshow.h
@@ -7,6 +7,6 @@ namespace Audioshow
     void loop();
     bool jack_plugged_in();
 
-    extern uint8_t leftPeak;
-    extern uint8_t rightPeak;
+    extern float leftPeak;
+    extern float rightPeak;
 }; // namespace Audioshow
